lastSentTime should check for both EC and non-EC
I believe it is possible to miss problematic situations (missing data)
when the EC encoder is sending data but the non-EC encoder for some
reason is not. Obviously it would be better to understand when that ever
happens and fix it, but it's better to at least report it than to ignore
it. This looks for both for setting lasSentTime, which is used to give
warnings.